requirements.txt
python-dotenv==0.10.1
twilio==6.14.0
requests==2.18.4
wget==3.2
